Temporarily add 'react-router-component' as dep
The components that depend on it should be either moved out or refactored to not need the router.  CollectionFilter could be rewritten to not use the router since its just switching views based on state/localStoreage. DefaultPath can use our NavigableMixin and FilterBar can use NetableLink or a Link we create.