NTI-1974 - FireFox sucks
We have to filter the inputs for FireFox. (which we were already doing)
However, we also have to know about the number field's 'min' value... so that
we do not allow the user to type a negative sign if the min is not positive.
(undefined, and negative un-restrict the key)