NTI-2484 - Actually, register the dialog to the modal manager.
The reverted change just listened to the manager for changes, and added/removed the hidden class... These dialogs cannot be over-lapped and can never be behind another. (Only one can be presented at a time, new ones forcibly destroy the previous if still open) So the only thing that this component needs to do, is register its instance with the manager so that Modals (not Dialogs) are masked.